Partial dump while DB is indexing

+++ NHerveToolboxMain/src/plugins/nherve/toolbox/image/db/DatabaseManager.java
@@ -74,12 +74,12 @@ public class DatabaseManager extends Algorithm {
 		return db;
 	}
 
-	public void index(final ImageDatabase db, final IndexingConfiguration conf, final boolean partialDump, final int waitMinutesBetweenEachDump) {
+	public void index(final ImageDatabase db, final IndexingConfiguration conf, final boolean partialDump, final double waitMinutesBetweenEachDump) {
 		db.clearDescriptors();
 
 		ImageDatabaseIndexer idxr = new ImageDatabaseIndexer(db);
 		idxr.setDoPartialDump(partialDump);
-		idxr.setPartialDumpSleep(waitMinutesBetweenEachDump * 60 * 1000);
+		idxr.setPartialDumpSleep((long)(waitMinutesBetweenEachDump * 60 * 1000));
 		idxr.setLogEnabled(isLogEnabled());
 		conf.populate(idxr);
